requirements.txt
google_auth_oauthlib==0.4.1
discord.py==1.3.2
discord==1.0.1
google_api_python_client==1.8.0
protobuf==3.11.3
python-dotenv==0.12.0
